rename: BookShelfPreview 컴포넌트 파일 명 오타 수정
